Changes for building in Android tree
This adds the Android.mk files and other miscellaneous changes necessary
for building smali/baksmali in the Android tree.

The ANTLR-based parser for smali is automatically generated from the ANTLR
parser definitions.

However, the jflex lexer needs to be manually generated using the runjflex.sh
script. The sources for the jflex lexer will be checked into the repository
whenever the lexer is updated and and the sources are regenerated.

Change-Id: I468f78f934e84daaeb86496110d614fe2d62b62b
Signed-off-by: Ben Gruver <bgruv@google.com>